Phase 2: Extract Multi-Timeframe Logic to dedicated service
Create MultiTimeframeCoordinator service to handle complex multi-timeframe operations:

- Extract load_multi_timeframe_data() method (244 lines)
- Extract _find_common_data_coverage() method (127 lines)
- Create MultiTimeframeCoordinator service in ktrdr/data/services/
- Replace DataManager methods with clean delegation
- Preserve all functionality and error handling behavior

Architecture improvement: DataManager now provides single-timeframe primitives
while MultiTimeframeCoordinator handles temporal coordination complexity.

Line reduction: 2,045 → 1,731 lines (-314 lines)
Functionality preserved: All unit tests pass (1098 passed, 1 skipped)
