+++ ktrdr/data/services/__init__.py
@@ -1,0 +1,6 @@
+"""
+Data services module.
+
+This module contains high-level data coordination services that orchestrate
+complex operations using DataManager primitives.
+"""
